refactor: generalize get/set frameStyle, update and  _buildStyle()
