+++ src/renderer/Frame.ts
@@ -1,3 +1,5 @@
+import { format } from "./Util"
+
 export interface BaseFrameOptions {
     darkMode: boolean
     minimizable: boolean
@@ -20,14 +22,53 @@ interface frameColors {
 
 type frameStyle = "windows" | "macos"
 
-export class Frame {
+export abstract class Frame {
     options!: BaseFrameOptions
     frame!: HTMLDivElement
 
     constructor() { }
 
+    abstract _build(): void
+
+    _buildStyle() {
+        const { colors = {} } = this.options
+
+        const colorsArray = Object.entries(colors)
+        const properties = colorsArray.map(([key, value]) => `--${format(key)} : ${value} !important`).join(';')
+
+        return properties
+    }
+
     remove() {
         this.frame.remove()
+    }
+
+    update() {
+        const hasFrame = Array.from(document.body.childNodes).includes(this.frame)
+        if (hasFrame) {
+            this.remove()
+        }
+
+        this._build()
+
+        if (hasFrame) {
+            document.body.appendChild(this.frame)
+        }
+    }
+
+    setFrameStyle(frameStyle: "windows" | "macos") {
+        if (frameStyle !== this.options.frameStyle) {
+            this.options.frameStyle = frameStyle
+            this.update()
+        }
+    }
+
+    set frameStyle(frameStyle: frameStyle) {
+        this.setFrameStyle(frameStyle)
+    }
+
+    get frameStyle() {
+        return this.options.frameStyle
     }
 
     get closeable() {
